Move getNip5PubKey() into nr-common again.
This is a repeat of the same change in 5a9a8c5a1a22c52d4cbdcda77f8203298748a259, because there were merge conflicts in `nr-common/` after @chmac reorganised the code in there.
